Revert "first attempt at re-doing removal of '/annotated_vcf' endpoint"
This reverts commit e3b3a0b9fc986a8049d420ce64d3ae4a5ac5777f.

+++ src/anyvar/queueing/celery_worker.py
@@ -212,13 +212,27 @@ def annotate_vcf(
         # annotate vcf with VRS IDs
         anyvar_app = get_anyvar_app()
         registrar = VcfRegistrar(anyvar_app.translator.dp, av=anyvar_app)
-        registrar.annotate(
-            Path(input_file_path),
-            Path(output_file_path),
-            compute_for_ref=for_ref,
-            assembly=assembly,
-            vrs_attributes=add_vrs_attributes,
-        )
+        already_annotated: bool = False
+        try:
+            registrar.annotate(
+                Path(input_file_path),
+                Path(output_file_path),
+                compute_for_ref=for_ref,
+                assembly=assembly,
+                vrs_attributes=add_vrs_attributes,
+            )
+        except ValueError as e:
+            already_annotated = str(e) == "Header already exists for id=VRS_Allele_IDs"
+            if already_annotated:
+                registrar.annotate(
+                    input_vcf_path=Path(input_file_path),
+                    output_vcf_path=None,
+                    compute_for_ref=for_ref,
+                    assembly=assembly,
+                    vrs_attributes=add_vrs_attributes,
+                )
+            else:
+                raise
         elapsed = datetime.datetime.now(tz=datetime.UTC) - task_start
         _logger.info(
             "%s - annotation completed in %s seconds", self.request.id, elapsed.seconds
@@ -239,6 +253,9 @@ def annotate_vcf(
                 elapsed.seconds,
             )
 
+        if already_annotated:
+            # return output file path
+            return input_file_path
         # remove input file
         Path(input_file_path).unlink()
         # return output file path
